+++ pkg/network/anonymity/settings.go
@@ -10,20 +10,20 @@ var (
 
 type SSettings sSettings
 type sSettings struct {
-	FServiceName   string
-	FF2FDisabled   bool
-	FRetryEnqueue  uint64
-	FNetworkMask   uint64
-	FFetchTimeWait time.Duration
+	FServiceName  string
+	FF2FDisabled  bool
+	FRetryEnqueue uint64
+	FNetworkMask  uint64
+	FFetchTimeout time.Duration
 }
 
 func NewSettings(pSett *SSettings) ISettings {
 	return (&sSettings{
-		FServiceName:   pSett.FServiceName,
-		FF2FDisabled:   pSett.FF2FDisabled,
-		FRetryEnqueue:  pSett.FRetryEnqueue,
-		FNetworkMask:   pSett.FNetworkMask,
-		FFetchTimeWait: pSett.FFetchTimeWait,
+		FServiceName:  pSett.FServiceName,
+		FF2FDisabled:  pSett.FF2FDisabled,
+		FRetryEnqueue: pSett.FRetryEnqueue,
+		FNetworkMask:  pSett.FNetworkMask,
+		FFetchTimeout: pSett.FFetchTimeout,
 	}).mustNotNull()
 }
 
@@ -34,8 +34,8 @@ func (p *sSettings) mustNotNull() ISettings {
 	if p.FNetworkMask == 0 {
 		panic(`p.FNetworkMask == 0`)
 	}
-	if p.FFetchTimeWait == 0 {
-		panic(`p.FTimeWait == 0`)
+	if p.FFetchTimeout == 0 {
+		panic(`p.FFetchTimeout == 0`)
 	}
 	return p
 }
@@ -48,8 +48,8 @@ func (p *sSettings) GetServiceName() string {
 	return p.FServiceName
 }
 
-func (p *sSettings) GetFetchTimeWait() time.Duration {
-	return p.FFetchTimeWait
+func (p *sSettings) GetFetchTimeout() time.Duration {
+	return p.FFetchTimeout
 }
 
 func (p *sSettings) GetNetworkMask() uint64 {